moved faction rules up 1 directory, split base Faction Mod into it's own file peace river faction mods now have specific class that extends base FactionModification class
