+++ lib/models/rules/nucoal.dart
@@ -1,20 +1,20 @@
 import 'package:gearforce/models/combatGroups/combat_group.dart';
 import 'package:gearforce/models/factions/faction_type.dart';
-import 'package:gearforce/models/mods/factions/peace_river.dart';
+import 'package:gearforce/models/mods/factions/faction_mod.dart';
 import 'package:gearforce/models/rules/rule_set.dart';
 import 'package:gearforce/models/unit/role.dart';
 import 'package:gearforce/models/unit/unit.dart';
 import 'package:gearforce/models/unit/unit_core.dart';
 
-class CEF extends RuleSet {
-  const CEF(super.data);
+class Nucoal extends RuleSet {
+  const Nucoal(super.data);
 
   @override
   List<UnitCore> availableUnits({
     List<RoleType?>? role,
     List<String>? filters,
   }) {
-    return data.unitList(FactionType.CEF, role: role, filters: filters);
+    return data.unitList(FactionType.NuCoal, role: role, filters: filters);
   }
 
   @override
