+++ lib/models/rules/rule_set.dart
@@ -1,7 +1,7 @@
 import 'package:gearforce/data/data.dart';
 import 'package:gearforce/models/combatGroups/combat_group.dart';
 import 'package:gearforce/models/factions/faction_type.dart';
-import 'package:gearforce/models/mods/factions/peace_river.dart';
+import 'package:gearforce/models/mods/factions/faction_mod.dart';
 import 'package:gearforce/models/roster/roster.dart';
 import 'package:gearforce/models/unit/role.dart';
 import 'package:gearforce/models/unit/unit.dart';
@@ -33,7 +33,7 @@ abstract class RuleSet {
 }
 
 class DefaultRuleSet extends RuleSet {
-  DefaultRuleSet(super.data);
+  const DefaultRuleSet(super.data);
 
   @override
   List<FactionModification> availableFactionMods(CombatGroup cg, Unit u) {
